Do not refresh seeds for every instance (#254)
* Call seed refreshing only once per event

Technically it is still called up to twice, depending on the value of
AllowToModifyEvent Marlin configuration, but that is not something we
can solve from here.

* Use actual type and remove thread_local

Given that these algorithms are marked as non-reentrant, I don't think
there is a way where they will ever run on multiple threads
concurrently.